Remove internal field from links.
In this crawler model, it's hard to say what an internal
link is. Basically you're saying that what's "internal" is
"matched by the include/exclude rules". If a URL doesnt, it
isn't returned in the crawl.

The closest thing to matching this functionality in other
crawlers would be to include the "won't crawl" URLs in the
result list with a special "won't crawl" status.