Wait for fetch functions before merging results.
Previously depth was (temporarily) nondeterministic. There was
a possibility that a "level 2" node would finish crawling before
a "level 1" node, and have its resulting links merged with the graph
before the "higher" node's.

Now we await the completion of all fetch functions before merging
their resulting links into a single queue. This has the downside of
using (potentially) lots of RAM to store queues for large crawls. We'll
see how important that is in practice.