Incorporate config into Crawler object.
Previously there was a Crawler type and a Config type that included
information that would govern the behavior of a Crawler. A Config was
embedded in Crawler.

This unnecessarily complicated the API for the crawler package,
especially because some fields needed to be public so that configuration
files could be unmarshalled into Config objects. Now there is a function
that will return a crawler from an io.Reader with JSON configuration
data.