Simplify API of sitemap package.
The Crawl CLI was relying on the sitemap package to recursively generate
lists of URLs from sitemaps and sitemap indexes. Looking forward, it was
clear that the CLI package would want to be able to log requests for
sitemaps. If that process was happening in the sitemap package, it would
be completely opaque to the CLI.

So, the API of the sitemap package is now only responsible for parsing
or fetching a single sitemap (or index).