Simplify internal type definitions for package sitemap.
When writing sitemap, I had only a basic understanding of xml.Unmarshal.
I think it's fair to say I still do, but I did pick up on the ">" symbol
for field annotation. This meant I could do away with the sub-types I
was using to descend through the XML hierarchy. And so I have.